add: datastructures and config internal packages
