+++ internal/datastructures/datastructures.go
@@ -1,0 +1,6 @@
+package datastructures
+
+type AllowedItem struct {
+	Name string
+	Type string
+}
