+++ pkg/tester/tester.go
@@ -1,21 +1,10 @@
-package shortinette
+package tester
 
 import (
-    "os/exec"
-    "bytes"
-    "log"
+	"fmt"
 )
 
-func RunUserCode(scriptPath string) (string, error) {
-    cmd := exec.Command("python", scriptPath)
-    var out bytes.Buffer
-    var stderr bytes.Buffer
-    cmd.Stdout = &out
-    cmd.Stderr = &stderr
-    err := cmd.Run()
-    if err != nil {
-        log.Fatalf("cmd.Run() failed with %s\n", err)
-        return "", err
-    }
-    return out.String(), nil
+func Run() error {
+	fmt.Println("Not Implemented")
+	return nil
 }
